# Do restat calculations during deserialization